Update and rename test_duchdns.py to test_duckdns.py
