+++ sewer/dns_providers/tests/test_duckdns.py
@@ -1,0 +1,66 @@
+import mock
+import json
+from unittest import TestCase
+
+import sewer
+
+from . import test_utils
+
+
+class Testacmedns(TestCase):
+    """
+    """
+
+    def setUp(self):
+        self.domain_name = "example.com"
+        self.domain_dns_value = "mock-domain_dns_value"
+        self.duckdns_API_KEY = "mock-api-key"
+        self.duckdns_API_BASE_URL = "https://some-mock-url.com"
+
+        with mock.patch("requests.get") as mock_requests_get:
+            mock_requests_get.return_value = test_utils.MockResponse()
+            self.dns_class = sewer.DuckDNSDns(
+                duckdns_token=self.duckdns_API_KEY,
+                DUCKDNS_API_BASE_URL=self.duckdns_API_BASE_URL,
+            )
+
+    def tearDown(self):
+        pass
+
+    def test_duckdns_is_called_by_create_dns_record(self):
+        with mock.patch("requests.get") as mock_requests_get, mock.patch(
+            "sewer.DuckDNSDns.delete_dns_record"
+        ) as mock_delete_dns_record, mock.patch("dns.resolver.Resolver.query") as mock_dns_resolver:
+            mock_requests_post.return_value = (
+                mock_delete_dns_record.return_value
+            ) = test_utils.MockResponse()
+            mock_dns_resolver.return_value = test_utils.MockDnsResolver()
+            self.dns_class.create_dns_record(
+                domain_name=self.domain_name, domain_dns_value=self.domain_dns_value
+            )
+
+            expected = {
+                "data": '{"domains": "example.com", "token": "mock-api-key", "txt": "mock-domain_dns_value"}',
+            }
+
+            self.assertDictEqual(
+                json.loads(expected["data"]), mock_requests_post.call_args[1]["json"]
+            )
+
+    def test_duckdns_is_not_called_by_delete_dns_record(self):
+        with mock.patch("requests.get") as mock_requests_get:
+            mock_requests_get.return_value = test_utils.MockResponse()
+            self.dns_class.delete_dns_record(
+                domain_name=self.domain_name, domain_dns_value=self.domain_dns_value
+            )
+            self.assertFalse(mock_requests_post.called)
+
+    def test_duckdns_is_called_by_delete_dns_record(self):
+        with mock.patch("requests.get") as mock_requests_get, mock.patch("requests.delete") as mock_requests_delete:
+            mock_requests_get.return_value = test_utils.MockResponse()
+
+            self.dns_class.delete_dns_record(
+                domain_name="example.com",
+                domain_dns_value="mock-domain_dns_value",
+            )
+            self.assertTrue(mock_requests_get.called)
